requirements.txt
python-dotenv>=1.0.0
pymysql>=1.1.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langsmith>=0.1.0
streamlit>=1.33.0
numpy>=1.26.0
scipy>=1.10.0
faiss-cpu>=1.7.4
pyyaml>=6.0
